// src/client/notebooksRoutes.ts
export interface NotebookParams {
  name?: string
  orgID?: string
  spec?: any
}
// ...
export interface PostNotebookParams {
  data: NotebookParams
}
// ...
export interface PatchNotebookParams {
  id: string

  data: NotebookParams
}
// ...
export interface DeleteNotebookParams {
  id: string
}